refactoring: Query를 사용하여 통계 처리를 DB로 넘김

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -147,11 +147,11 @@ class BookServiceTest @Autowired constructor(
         val results = bookService.getBookStatistics()
 
         // then
-        assertCount(results, BookType.COMPUTER, 2)
-        assertCount(results, BookType.SCIENCE, 1)
+        assertCount(results, BookType.COMPUTER, 2L)
+        assertCount(results, BookType.SCIENCE, 1L)
     }
 
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(results.first { it.type == type }.count).isEqualTo(count)
     }
 }
